requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
unstructured>=0.12.0
libmagic>=1.0
python-magic>=0.4.27
tiktoken>=0.5.0
